clean renamer class
