+++ lib/thor_template.rb
@@ -6,4 +6,5 @@ module ThorTemplate
   autoload :Command, 'thor_template/command'
   autoload :Version, 'thor_template/version'
   autoload :Generator, 'thor_template/generator'
+  autoload :Renamer, 'thor_template/renamer'
 end
