+++ lib/thor_template/generator.rb
@@ -40,55 +40,7 @@ module ThorTemplate
     end
 
     def rename
-      Dir.chdir(@name) do
-        Dir.glob("**/*") do |path|
-          next unless File.file?(path)
-          rename_content(path)
-
-          next unless File.file?(path)
-          rename_path(path)
-        end # Dir.glob
-
-        remove_empty_directories
-      end
-    end
-
-    def rename_content(path)
-      content = IO.readlines(path)
-      result = content.map do |line|
-        line = line.gsub(/ThorTemplate/, @name.underscore.camelize)
-        line = line.gsub(/thor_template/, @name.underscore)
-        line = line.gsub("USER_PROVIDED_NAME", @name) # special case
-        line
-      end
-      IO.write(path, result.join(''))
-    end
-
-    def rename_path(src)
-      dest = if special_rename?(src)
-        src.gsub(/thor_template/, @name)
-      else
-        src.gsub(/thor_template/, @name.underscore)
-      end
-
-      folder = File.dirname(dest)
-      FileUtils.mkdir_p(folder) unless File.exist?(folder)
-      FileUtils.mv(src, dest) unless src == dest
-    end
-
-    def special_rename?(path)
-      %w[
-        thor_template.gemspec
-        bin/thor_template
-        lib/thor_template.rb
-      ].include?(path)
-    end
-
-    def remove_empty_directories
-      Dir['**/*'].
-        select { |d| File.directory? d }.
-        select { |d| (Dir.entries(d) - %w[ . .. ]).empty? }.
-        each   { |d| Dir.rmdir d }
+      Renamer.new(@name).rename!
     end
 
     def git
